+++ lib/rodf/compatibility.rb
@@ -1,7 +1,3 @@
-# Copyright (c) 2010-2012 Thiago Arrais
-#
-# This file is part of rODF.
-
 require 'builder'
 
 if !String.method_defined? :to_xs