+++ lib/rodf/text.rb
@@ -1,7 +1,3 @@
-# Copyright (c) 2010 Thiago Arrais
-#
-# This file is part of rODF.
-
 require 'builder'
 
 require_relative 'document'
@@ -48,4 +44,3 @@ module RODF
     end
   end
 end
-