+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## Unreleased
+- Nothing yet
+
+## v1.1.1
 - [Issue #36](https://github.com/westonganger/rodf/issues/36) - Improve compatibility for Float Cells
 - Fix gemspec require for Ruby 1.9
 - Cleanup File Commenting